+++ packages/app-connector/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @parcnet-js/app-connector
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @parcnet-js/podspec@0.0.3
+  - @parcnet-js/client-rpc@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes