+++ packages/client-rpc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @parcnet-js/client-rpc
 
+## 0.0.4
+
+### Patch Changes
+
+- Support metadata on PodSpec schemas
+- Updated dependencies
+  - @parcnet-js/podspec@0.0.3
+
 ## 0.0.3
 
 ### Patch Changes